feat: add hook to get max items by window size
