+++ src/hooks/useIsMobile.js
@@ -1,0 +1,20 @@
+import { useState, useEffect } from 'react';
+
+function useIsMobile(breakpoint = 743) {
+  const [isMobile, setIsMobile] = useState(window.innerWidth <= breakpoint);
+
+  useEffect(() => {
+    const handleResize = () => {
+      setIsMobile(window.innerWidth <= breakpoint);
+    };
+
+    window.addEventListener('resize', handleResize);
+    handleResize(); 
+
+    return () => window.removeEventListener('resize', handleResize);
+  }, [breakpoint]);
+
+  return isMobile;
+}
+
+export default useIsMobile;
